Creating MapDef to hold maps definitions.

+++ src/definitions.h
@@ -7,7 +7,7 @@
 #define TILE_SIZE 10
 
 typedef struct Map {
-	u16 *tiles;
+	u8 *tiles;
 	u32 size;
 	u32 width;
 	RangeInt steps_to_combat;
